UI: Added theme menu
