+++ ChangeLog
@@ -1,5 +1,6 @@
 2020-06-14 (12.19)
 	UI: Use theme colours in main display #94
+	UI: Added theme menu
 
 2020-06-14 (12.19)
 	FLTK: line numbers colours #93
